Implemented Router/Service/Controller style for canvar router. Also implemented on syllabus and files endpoints - (10/19/24)
